CarpetIOHDF5: Use CCTK_REAL instead of double for meta data
Use the type CCTK_REAL instead of double for storing meta data in the
HDF5 files.  This is necessary if CCTK_REAL has more precision than
double.

darcs-hash:20050411170627-891bb-374e4c2581155d825f9a1925b1d4319051bc36d6.gz

+++ Carpet/CarpetIOHDF5/src/Output.cc
@@ -938,7 +938,7 @@ static int WarnAboutDeprecatedParameters (void)
   {
     CCTK_WARN (2, "Parameter 'IOHDF5::out3D_dt' is deprecated, please use "
                   "'IOHDF5::out_dt' instead");
-    snprintf (buffer, sizeof (buffer), "%f", out3D_dt);
+    snprintf (buffer, sizeof (buffer), "%f", (double)out3D_dt);
     CCTK_ParameterSet ("out_dt", CCTK_THORNSTRING, buffer);
     warnings++;
   }
